requirements.txt
openai
python-dotenv
yfinance
pandas
numpy
datetime
typing
fpdf2
jinja2 
weasyprint
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
beautifulsoup4
deepl